+++ pkg/barback/CHANGELOG.md
@@ -6,6 +6,9 @@
   declared. This is valid for transformers like dart2js that need to read their
   primary input in order to determine whether they should run.
 
+* Fix a deadlock bug when a lazy primary input to a lazy transformer became
+  dirty while the transformer's `apply` method was running.
+
 ## 0.13.0
 
 * `Transformer.isPrimary` now takes an `AssetId` rather than an `Asset`.